# Conflicts:
#	Visual Studio/RobotArminator/RobotArminator/RobotArminator.cpp

+++ Visual Studio/RobotArminator/RobotArminator/VisionPosition.hpp
@@ -10,11 +10,11 @@ namespace RobotArminator
     {
     public:
         VisionPosition();
-        VisionPosition(float X, float Y, Time time, Orientation orientation);
+        VisionPosition(double X, double Y, Time time, Orientation orientation);
         ~VisionPosition();
 		bool isDefault;
-        float X;
-        float Y;
+		double X;
+		double Y;
         Time time;
         Orientation orientation;
     };
